Moved dataType functions into cpp file to avoid compilation errors
